Restructure code into two folders for the jobs of building workshops or upgrading blueprints (#16)
