+++ src/buildWorkshop/importEventProducts.ts
