+++ test/buildWorkshop/computeIdealLevelsForEvent.test.ts
