+++ test/buildWorkshop/productLooper.test.ts
